Account cache in McAccountManager and call_when_ready_with_accounts
Implement an cache (as GHashTable) of McAccounts into McAccountManager.
Add mc_account_manager_get_account() to get an account from the cache,
instantiating a new object in case it was not there.
Add mc_account_manager_call_when_ready_with_accounts(), which calls a callback
when the account manager and all the known accounts are ready.

git-svn-id: https://projects.maemo.org/svn/chavo/trunk/framework/open/telepathy-mission-control@20535 64d1ce6a-1406-0410-9ac8-afed03b27183