Renaming of variables and changing access rights of properties
In McdChannel, deprecate the properties "channel-<property>" in favor of
"<property>"; introduce the new properties as writable construct only
properties, as we have accessor methods for them.
Rewrite the existing code not to use g_object{set,get} on the new properties.
Rename some private variables to strip off the "channel_" prefix.