Rewrite error handling on channel requests
It doesn't seem right that errors occurred while creating accounts are emitted
on the channel dispatcher as dispatch errors. So, this patch moves handling of
these errors to the code in McdAccount which creates the requests, by having
the error stored in the McdChannel and inspected when the channel status
changes to FAILED.
Then the error on the dispatcher is issued from mcd-account.c; it's still
wrong, but we need it if we want to behave as MC4 did.