Dispatching on NewChannels signal
Requested channels are no longer dispatched when the CreateChannel method
returns, but instead we wait for the NewChannels signal (because there might be
other channels coming along, and we try to dispatch them all together).

Move the logic for determining whether a channel must be processed by MC into a
separate function.