Telepathy.Client interface: Get the *ChannelFilter properties according to the spec
Look for Telepathy.Client on the bus instead of the .client files as a
canonical source and parse the .client file only if it exists as an
optimisation.

ListNames and ListActivatableNames are called. If the name start with
o.f.T.Client, we create the McdClient struct and fill it with either the
.client file if it exists or with D-Bus calls to get the Interfaces property
and the *ChannelFilter properties

HandlerChannelFilter property is saved in McdClient->caps and used to call
SetSelfCapabilities.