Reinvoke the channel handler with the correct request and user time.
The channel handler must be invoked on the request McdChannel object, so that
the dispatcher will extract the request object path and user action time from
it.
Make sure the channel request is disposed after it has succeeded (if we already
have a McdChannel for that tp channel).
Move the code to McdDispatcher, it seems more fit there.