Keep the channel alive in the online requests queue
When putting the channel into the online_requests queue, reference it; the
unreference will be done when the callback will be invoked.

This is needed to avoid a crash that can happen if the channel is aborted while
we are waiting for the connection; it's probably not triggered now, but it will
happen when channel request cancellation is implemented.