Emit the request Failed signal as last.
The Failed signal causes the client (using libmcclient) to delete all
references to a request; if a RemoveFailedRequest is received after that, it
cannot be mapped anymore. This change makes it so that the client will receive
RemoveFailedRequest before the Failed one, so that the client's request ID is
still valid.