Dispatch requested channels waiting for approval
If an incoming channel is waiting for approval and a client requests the same
channel (via EnsureChannel), then the DispatchOperation must finish
immediately.
This patch makes it so that if the channel is found in a dispatch context and
is waiting for approval, HandleWith() is called on the dispatch operation, so
the channel handler will eventually be invoked.