Check for autoconnection conditions
Add _mcd_account_maybe_autoconnect(), which checks if the account could go
online, and attempt connecting.
This function can be called anytime, and especially when there are some events
that could open the possibility of connecting an account.

Here, we call it when the account has been loaded, and when a parameter changes.