Implement channel recovery
While examining the channels already present on the TpConnection, recover those
which don't match any known channels.
Recovering is done by querying the running handlers about the channels they are
already handling, and mathing them with the ones to be recovered: if a channel
is already being handled, nothing has to be done; if instead a channel is not
handled by any clients, this channel has to be redispatched.