Fix channel direction
Channels requested not through MC were having their direction always set to
incoming; this is because after the "tp-proxy" property is set, GObject was
setting the "outgoing" property to its default value, therefore overriding
whatever had been set before.
The solution is to hold on the call to mcd_channel_setup() until the object is
constructed.