_mcd_account_connect, _mcd_connection_connect: don't consume @params
Instead of passing parameters which are consumed by the function, it seems
safer and more comprehensible to use GHashTable reference counting.

Also, instead of stashing the params on the McdConnection with qdata,
put them in a private struct field (more legible *and* more efficient).

I believe that this branch also fixes a memory leak in
mcd_account_connection_proceed: if plugins refused connection to the
account, the params hash table would be leaked.