_mcd_connection_connect: temporarily duplicate this function
This function has two entirely distinct uses: it's called with
params == NULL from within McdConnection when it wants to trigger a
connection attempt (which pokes the McdAccount, which eventually calls
into it again as the second form), and it's called with params != NULL
from within McdAccount. This is needlessly confusing, so split them up.