Keep the context alive
Don't let the context be freed in the middle of a function.
This was causing a crash, if the handler returned an error: the channel was
aborted, therefore removed from the DispatchOperation, which in turn emitted
the Finished signal. This caused the mcd_dispatcher_context_handler_done() to
be called and free the context, which would still be used later.