McdConnection: ignore NewChannel until we know whether we have Requests
Before Requests became mandatory, this was a minor optimization and a
questionable correctness fix: MC could remember whether legacy channels
had Suppress_Handler or not, and avoided calling ListChannels.

Not calling ListChannels avoided accidentally dispatching old channels
which in fact had Suppress_Handler (before Requests, this information
couldn't be recovered) but potentially meant that we failed to dispatch
old undispatched channels that did not have Suppress_Handler, so we can't
really win.

Now that we expect all CMs to implement Requests, this code is just
confusing. Getting rid of it means that channels which appear before
the TpConnection is ready will cause crash-recovery code to be invoked,
altering the behaviour of one of the tests, so I've adjusted the test to
reflect this.

This does mean we'll potentially re-dispatch channels that are already
being handled by a handler that is not a Client, but as noted above, with
pre-Requests CMs we can't win anyway.