McdDispatcher: sort possible handlers by "quality"
Quality is currently defined as: handlers that bypass approval are better
than handlers that don't; otherwise, handlers get one point for each
matching property in a filter, and the handler with most points is best.

In practice this should mean that the most specific bypass-approval filter
(if any), or the most specific filter (otherwise), wins.