McdDispatcher: rethink logic for selecting and running handlers
The previous logic would keep dispatching subsets of the batch of channels
until no more could be dispatched. However, this is not actually what we
want.

By this point we've already decided what the preferred handlers are, in
preference order, and reassured ourselves that each possible handler in
the list can handle all of the batch of channels. Also, filters are
immutable, so they can *still* all handle all the channels. So, the right
thing to do seems to be to pick the first one that still exists, and
dispatch everything to it.

The only failure mode is that we decided in advance what the possible
handlers were, but they've all exited, and none of them are activatable.
This means we can no longer handle all these channels as a unit at all.
For now, recover by closing all the channels (in future, it might be worth
exploding the batch into individual channels and launching one handler
each, or exploding the batch into individual CDOs and asking the
approvers again, or something).

(Also, this patch is a net code reduction.)