Rename non-standard interfaces into the com.nokia namespace
The Creation interface remains .DRAFT, but since MC beta 70 and
telepathy-spec 0.17.24 the same functionality has been present on the
main CreateAccount method. Any existing users should migrate to the core
API, at which point the draft API can be deleted altogether.

ChannelRequests won't go into telepathy-spec, we already have API for
this on the ChannelDispatcher object.

Compat is for compatibility with MC 4, which was never specified by
telepathy-spec, so it would be inappropriate for it to go in
telepathy-spec.