tests: stop asserting that MC will call Get(HANDLER, "HandledChannels")
Previously, we asserted that it would be called. Now, we allow it to be
called, and return the right thing, but don't fail if this doesn't happen.

This means we no longer have precise control over the order of events in
dispatcher/create-at-startup.py, so we can't assert that ObserveChannels
for the request and the incoming channel are called in the right order.
Temporarily disable that assertion.