Don't close channels on dispose
There are various reasons why MCDChannels are disposed, on the current desktop
this is either when a client calls close on the telepathy channel directly or
when disconnecting from a connection. In the first case calling Close makes no
sense (somebody already did this), in the second case it causes troubles when
Close has an extra protocol level meaning. For example when using an IRC proxy,
closing a the group text means parting the channel on irc, this means on
disconnect your proxy will no longer be part of the cahnnel. Which is exactly
what you didn't want :)