McdDispatcher: fix the McdDispatcherContext reference-leak
We never release all the client locks if there are no channels left. As
a result, if some client locks are held (i.e. we have taken a reference
to the context that will be released when all locks are released), we need
to release the reference at the point that we decide not to release the
locks.