Add a configure option for whether to support plugins
In Debian we don't want to support MC's plugin architecture yet (if at
all), to reduce the amount of API/ABI exposed.

When plugins are disabled, we don't need GModule,
libmissioncontrol-server doesn't need to be installed or be a shared
library, and the headers don't need to be installed either.