mcd_account_request_presence_int: allow same presence to be re-requested
Also, only emit AccountPropertyChanged for RequestedPresence if it
actually changed, and add a regression test for both.

Transport plugins provide a hook for MC to communicate with connectivity
mechanisms like NetworkManager, Maemo's ConIc or Moblin's ConnMan to
request an Internet connection when signing on to IM protocols.

If the user sets online presence but the Internet connection fails, the
RequestedPresence will still be online. As a result, the user can only
retry connection by setting their presence to offline and back to online,
which is deeply unintuitive.

Removing the check for a no-op means that the UI re-requesting an online
presence will trigger another attempt to connect to the Internet; this one
might succeed.

Also, if a protocol has side-effects from re-requesting the same presence
(e.g. it records the time at which the current presence was requested),
this change will make it behave according to the user's input.