fd.o #21299: don't reconnect if disconnected with reason Name_In_Use
On protocols where the new connection fails, we'd just keep retrying
until the old connection goes away (i.e. perhaps indefinitely) which is
not too bad, but a waste of bandwidth.

On protocols where the old connection drops and the new one succeeds,
we could easily get into a fight with the other client - we reconnect, the
other client gets kicked off, it reconnects, we get kicked off, we
reconnect, repeat. This is particularly harmful when our initial
reconnection timeout is only a second. If the other client is also
MC 5, we could end up with two connection attempts per second,
indefinitely.