McdAccount: remove incorrect "optimization" using an array of property GValues
Previously, this code would copy changed properties into a GArray of
GValue, populate a GHashTable with pointers into the GArray's memory
buffer, and emit that GHashTable in the delayed PropertiesChanged signal.

This was fine until the GArray ran out of space in its memory buffer
and triggered a realloc(), at which point the GValues would be copied
to the new buffer, leaving the values in the GHashTable pointing to the
old (cleared and/or freed) memory buffer. At this point, emitting the
signal would cause a crash. The regression test
account-manager/make-valid.py causes this failure mode.

In practice, slice-allocating the GValues individually is likely to be
just as efficient, so do that instead.