McdAccount: when forcing loaded AutomaticPresence to be online, set status member too
Alberto points out that the status is used (if supported) regardless of
the value of the presence type, with the presence type used only to select
fallbacks.

This is mitigated by the fact that if we fail to set the AutomaticPresence
(because it's not an online presence type), we'll keep the default
available presence that was obtained by signing in.