McdChannel: never implicitly close the channel
This hurts clarity, and is actually actively undesirable. We now close the
channel explicitly in every situation where we want to, so the only
remaining times that we abort channels are they are closed, or when the
McdConnection aborts (i.e. the TpConnection closes). In either case, there
is no need to close the channel, because the CM has closed it or will do
so shortly.

The undesirability comes from the fact that in telepathy-idle, closing a
room Channel is like a /part, but disconnecting the Connection is like a
/quit. When Idle is connected to an IRC proxy like irssi or bip, we never
want to /part from channels in response to disconnection from the proxy,
because that would mean the proxy would no longer stay in your desired
channels recording a backlog.

(See commit 6cda49f79b43a8a71 for a more targeted form of this change,
as applied to Mission Control 4 upstream and in Debian.)