McdDispatcher: refactor how plugins close channels
mcd_dispatcher_context_process has two purposes: it releases the filter's
implicit reference to the context/lock on the context, and it can
arrange for channels to be terminated.

It seems cleaner to introduce mcd_dispatcher_context_proceed() which
just releases the reference/lock at all times, and redefine
mcd_dispatcher_context_process (which will be deprecated later) in terms
of it.

To ensure that proceeding after channels have been "dealt with" does not
invoke clients, new function mcd_dispatcher_context_forget_all provides
a way to tell the context that its channels are no longer suitable for
dispatching, removing them (as a result of the abort signal). proceed
needs to handle this by just clearing up if all the channels have already
gone.

The remaining bit of new API needed is an easy way to terminate channels,
possibly with a reason or with Destroy. mcd_dispatcher_context_close_all
and mcd_dispatcher_context_destroy_all provide this.