fd.o #22585: src: stop generating interfaces and GTypes, fixing static linking
All the interfaces and GTypes we need are already in libmcclient, with the
same names and auto-generated implementations, so it's useless to generate
them in libmissioncontrol-server.

When configured with --enable-static --disable-shared (which is what we
want in Debian, I suspect), it becomes actively harmful to have two
implementations: shared library linking can survive colliding symbols
(at least on Linux), but static linking can't.

(If we ever want Mission Control to be independent of libmcclient, then
it should generate its own quark/type functions in the Mcd namespace.)