McdConnection: if no fallbacks for a presence are supported, twice, don't crash
The first time this happened, we'd overwrite the NULL at the end of
one of the fallbacks arrays (e.g. _away_fb[1]) with a pointer to
"available". Then, the *second* time it happened, we'd iterate off the end
of the array (since it would no longer have a NULL terminator), try to
dereference some random memory as a char* in g_str_hash, and crash.