McdConnection: before calling Connect, try to set initial presence
If this fails, it's no big deal - we'll set it again later anyway.

The infrastructure I've set up is fairly generic, so we can re-use it for
anything else that we want to try to set up before going online (to
include ContactCapabilities in 5.3, and perhaps Avatars and Aliasing
later).