account-manager/presence.py: don't assert about the ordering of SetPresence vs AccountPropertyChanged(CurrentPresence)
They might come out in either order, in fact (it's a race). The test seems
to be more likely to lose the race if MC is slow, for instance because
it's run under qemu (Scratchbox) or valgrind.

Also make the failure message clearer if things happen in the wrong order.