McdDispatcher: disconnect from McdClientProxy signals whenever we drop a proxy
In particular, this guarantees that even if a client proxy outlives the
dispatcher, the dispatcher's callbacks won't be invoked when the client
finds out its unique name or becomes ready.

Each call to mcd_dispatcher_discard_client is annotated with an
explanation of why either it can't leak startup locks, or it doesn't matter
if it does.