McdDispatchOperation: rename ->finished to be less misleading, and correct result of _mcd_dispatch_operation_is_finished
The "finished" boolean didn't actually mean "we have emitted Finished",
it meant "we will emit Finished as soon as we are allowed to", so
wants_to_finish would be a more accurate name.

The only current caller of _mcd_dispatch_operation_is_finished is the
McdDispatcher, when listing dispatch operations. To be consistent with
the change-notification signals, this API should list all dispatch
operations that have not yet emitted the D-Bus signal, even if they are
going to emit it as soon as their clients have finished.