McdDispatcher: simplify HandleChannels invocation
We never attempt to dispatch subsets of the context's channels to a
handler, and the set of channels is "reduce-only", so there's no point in
remembering the channels so we can compute their intersection with those
remaining in the context - that intersection will always be equal to the
set of channels in the context.

This is just as well, since the way we tracked the subset of channels was
quite precarious (keeping pointers to them without reffing those pointers,
then checking whether they were still listed in the context).