_mcd_dispatcher_reinvoke_handler: reimplement with correct behaviour
While it's superficially similar to the end of the dispatching pipeline,
the requirements are actually rather different.

When we call HandleChannels the first time, we can use any handler,
errors must be handled by trying to dispatch to a different handler, and
we need to keep track of who's handling the channel at the end.

When an EnsureChannel() call causes us to reinvoke HandleChannels, the
previous implementation (using the dispatcher context) was buggy:

* if a "better" Handler has appeared since the channel was initially
  dispatched, it would dispatch to that one, resulting in the channel
  being handled by two different processes (!)
* if handling with the selected client failed, it would try others
* if handling failed with all clients, it would close the channel

I've corrected these as follows:

* only Handlers that match the unique name that we know is already
  handling the channel will be considered
* there is only one attempt to call HandleChannels
* if the attempt fails, or there is no suitable Handler object in the
  process any more, we act as though HandleChannels succeeded - the
  Handler won't be notified, so it can't bring itself to the foreground
  or whatever, but if it wanted that it should have kept its Handler
  object