McdDispatcher: only remember unfinished CDOs, and disconnect from their signals correctly
For now, "finished" CDOs should keep themselves referenced until they can
actually be freed (the definition of "finished" will be corrected later).

We only need to store CDOs that want approval in a list, and we only need
to do so until they emit finished: the only purpose of having this list,
apart from the OperationList interface, is so EnsureChannel requests can
cause approval.