McdClientRegistry: don't kick the daemon, it makes it forget pending activations
This is essentially a forward-port of commit c51127c9 from master.

D-Bus forgets about all current activations when it's asked to reload its
configuration. This causes bad side-effects if MissionControl was in the
process of activating itself (and for everything else that's activating itself
on system startup).

The upstream D-Bus bug is fd.o#24350, this also reopens fd.o#23921