McdAccount: guarantee to emit ConnectionStatus and ConnectionStatusReason in the same signal
Previously, we relied on the 10ms delay hack to get them to be in the
same signal, but this way seems better.

This requires a small expectation change in
account-manager/auto-connect.py: previously we were relying on the 10ms
delay hack delaying the emission of
{ ConnectionStatusChanged => CONNECTING } until after Connect() was
called, which doesn't really seem right in any case (conceptually, we're
trying to connect even before we call RequestConnection).

Similarly, avatar-persist and avatar-refresh encode the same faulty
assumption. Since that code is adequately tested in auto-connect.py,
and isn't really the point of the avatar tests, I've just removed the
faulty expectation from them.