Load and use the account storage plugins, with two pseudo-plugins by default.
The pseudo-plugins are the keyfile plugin (an oject that implements the
relevant plugin interface, even though it isn't loaded as a plugin) and
the gnome keyring plugin (ditto) which are loaded at advertised priorities.

When no other plugins are loaded, the keyring plugin should transparently
steal all secret parameters from accounts in the mcd-account-manager,
provided that the keyring daemon is accessible.