SSO accounts can have multiple services, but this is not exposed as a property.
Instead, expose it as a special property which is not covered by the standard
iterator but will be returned if you ask for it.

Also, exclude the property from being saved, since the list of services can
change and should not be cached in the SSO storage for the account itself.