Don't cache whether we have the requests interface
One use of the caching (in _mcd_connection_request_channel) was subtly
incorrect: before using the cached variable we checked whether the
TpConnection was ready, but that's not the same thing as checking whether
our on_connection_ready() callback (which caches it) has actually been
called yet.

This fixes a regression in dispatcher/connect-for-request.py.