Only flush changed account details to disk.
Previously every time any property of any account changed, every account
was flushed to storage. This is really, really bad if you use
gnome-keyring, because it meant a pair of calls to Unlock()/CreateItem()
for each secret on each account, every time anything changes. On my
machine, signing in 6 accounts entailed 384 pairs of calls — 64 per
account! Furthermore,
it seems that the implementation of CreateItem() calls fsync(), so my
hard disk would thrash for about thirty seconds.

This patch adds an optional argument to
mcd_account_manager_write_conf_async() and mcp_account_storage_commit(),
indicating which account changed. If NULL, everything is flushed; if
non-NULL, only that account is flushed. In my testing this reduced
logging in to 9 Unlock()/CreateItem() pairs, which is still more than
zero, but stops the disk thrashing quite so badly.

Removing accounts doesn't cause the secrets to be removed from the
keyring... but that's not a new bug. Everything else seems to work as
expected.

I didn't update the libaccounts-glib backend to benefit from this, but
given that all it does is iterate across all accounts and call two
methods on each, it should be easy enough.

Mitigates: http://bugs.freedesktop.org/show_bug.cgi?id=28390