fd.o #24645: _mcd_dispatcher_reinvoke_handler: try to use the same Handler
If a process has Handler well-known names A and B, and a channel
that was already passed to A's HandleChannels implementation is returned
by EnsureChannels, we want to call HandleChannels on A if it still exists,
or failing that, B if its filter matches the channel.

This is particularly useful if A was a PreferredHandler with an empty
list of filters, as used by the proposed
tp_account_create_channel_and_handle_async().