Retrieve the account storage plugin lazily.
Apparently mcd_account_manager_get_storage_plugin can be called too early when
the account is created, before it is written to the storage plugin.

This patch makes it reference the account storage plugin lazily when needed,
and to fail when it can't retrieve it.

I tried to have it retrieve the storage plugin consistently at a later stage,
but there is no way of knowing when the account is written to the storage
plugin since there are multiple code paths. So this seemed like the safest
choice.