Copy the TpChannel to a proxy McdChannel when setting it up
This fixes an assertion failure when emitting SucceededWithChannel for a
proxy (secondary) representation of a channel (i.e. the second and
subsequent EnsureChannel requests that returned a particular channel).

To do this without putting a non-primary representation of a channel
into the dispatching machinery (which breaks tracking which requests
where satisfied) it's necessary to add _mcd_channel_is_primary_for_path;
see that function's doc-comment for the definition of "primary" and
"proxy" in this context.