mcd-account-requests: defer Proceed() until the dispatcher has been told
This matches the order of events in the modern API, and means that the
implementation of Proceed() can safely assume that the dispatcher knows
about the channel request.

This does break the expectations of the old-API tests: previously, we
incorrectly asserted that AddRequest happened after calling CreateChannel
on the connection manager, but either order would be valid, and in fact
the new order makes more sense.