New McdStorage interface for storing/retrieving account settings + params
The McpAccountStorage interface used to communicate with the storage plugins

The McdStorage interface is what MC in general should use to store/inspect/
update/whatever account details, and is responsible for managing the storage
plugins. It replaces the old method of stealing the account manager's
internal data structures via the McdAccount and changing them behind
the account manager's back.