Make _mcd_account_dup_parameters() synchronous
This is a really massive patch, and I'm very sorry.

Previously _mcd_account_dup_parameters() took a callback parameter. This
was because mcd_account_get_parameter() took a callback parameter. The
latter used to actually be asychronous in the gnome-keyring case, but it
is not any more, thanks to the harmonious account backend API.

So we can remove a *lot* of unnecessary complexity if we make the latter
return the parameter's value synchronously, and hence make the former
return a GHashTable synchronously. In particular, this means we can make
retrieving the Parameters D-Bus property synchronous, which means we can
port to TpDBusPropertiesMixin.

I did try to split this patch up, but as you, dear reader, will be able
to see from what I ended up with, this code was all really entangled
with itself. This is the best I could do.