create-at-startup: don't expect ObserveChannels calls so close together
Now that preparing the connection takes one more step, it can take
longer before ObserveChannels is called with the channel that just
appeared on connecting, so let's allow that to appear earlier.

This only relaxes the test assumptions a little. Although it currently
fails on 0.13.9 and this fixes it, it'll still work on tp-glib <
0.13.9.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>