create-with-properties: remove unsound signal expectation
When this test creates a new account, it expects four events to occur:

• A.AccountPropertyChanged, to say the new account is valid;
• AM.AccountValidityChanged, to say the new account is valid;
• CreateAccount returning successfully;
• RequestConnection on the fake CM, because ConnectAutomatically was
  set to True and RequestedPresence was set to non-offline in
  CreateAccount.

However, if the account's not yet been announced by
AccountValidityChanged, MC has no reason to emit AccountPropertyChanged
on it. It happens that, in some cases, it does; but this depends on
whether or not MC has already introspected the fake CM by the time the
account is created. So the test is wrong to expect
AccountPropertyChanged. Instead, we check that Account.Valid is in sync
with what AccountValidityChanged said.

This fixes an issue where this test would only pass if it was run after
account-manager/create-twice, but not if it was run individually.