Accounts-GLib backend: fix shadowing warnings
acct() is a function defined in unistd.h, which something winds up
including in this file.

Sometimes 'acct' was an AgAccount, and other times it was the
"cm/protocol/foo" identifier. I changed the former to 'account' and the
latter to 'account_suffix' where needed, as a first step towards
coherency.