accounts-glib: List services in constructed, not init
I'm about to replace the hard-coded AG_SERVICE constant with a field in
the class (so that a subclass can look for other services and set Hidden
to True for those accounts). However, as we discovered way back in
f7b7aa8e477abe03c1340c2b200ddcf9c9792c2c, when _init runs for class T,
G_OBJECT_TYPE() always returns T even if we're constructing a subclass.
So we're going to need to wait until constructed() to get at the desired
service name.