UpdateParameters: apply Unset parameters immediately, if possible.
If a parameter in the Unset list was previously set to its default
value, there's no need to reconnect to apply the change. Also, if its
default value is different, but it's got the DBus_Property flag, we can
apply it to the running connection.

Distressingly, fixing this does not actually make this code much
simpler.

Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=34316>