UpdateParameters: break apart check_parameter_update()
I had thought that the logic for setting and clearing parameters was
similar enough that combining them would be a net win, but it seems that
that was not the case. This patch rearranges check_one_parameter()
check_parameter_update() into check_one_parameter_update() for parameter
updates, and check_one_parameter_unset() for unset parameters. There is
a little duplication, but these are much easier to follow.

Dear reviewer, I'm sorry for the many false starts in this branch. I
did try to rebase them out of existence, but…