Preserve deleted account while calling Identify
When an account is deleted, we'd expect that the McdAccount object will
die shortly afterwards. However, we need to ensure that RemoveAccount is
called on connection managers that implement AccountStorage. Previously,
the account object was used as the weak_object for the callback to
IdentifyAccount; this could mean that the callback would never fire, and
thus we'd never get around to calling RemoveAccount.

So this patch holds a reference for that callback.