Be more robust against cataclysmic test failures
Previously, if MC never claimed its bus names (because it was crashing
on startup, for instance), the exception in the test initialization code
wasn't caught. This means nothing would ever quit the Twisted mainloop,
and the (failed) test would sit there forever.

This patch ensures that the test exits uncleanly if this happens.