Tests: remove stale FIXMEs about closing channels
Since 5dfe2f0, Mission Control has not called Close() on channels when
you disconnect. The test for this incorrect behaviour was commented out
at some point, with a FIXME that the behaviour should be changed.

Well, now the behaviour has been changed. So I replaced one with an
assertion that Close() is not called, and the other I just removed
(since that test doesn't actually have any channels).