mcd_account_check_validity: report a GError for invalid accounts
This corrects the specific case of poor error reporting mentioned on
<https://bugs.freedesktop.org/show_bug.cgi?id=40337>, where MC
previously said “The supplied CM parameters were not valid” but now
actually says which one.

There are a couple of other account-creation cases (namely “CM doesn't
exist” and “CM doesn't have that protocol”) which don't give the correct
errors. And the test doesn't test the “Protocol doesn't have one of the
specified parameters” case.