Stop building docs for libmissioncontrol-server
I'm hitting a weird issue in gtk-doc where the handwritten .types file
is overwritten with a generated one during distcheck (but not during
normal out-of-tree builds). Given that we have a newer plugin API we're
keener on, let's just stop building documentation for the old one. (Most
distros do not distribute libmissioncontrol-server as a library anyway.)

The mission-control-plugins documentation doesn't hit this issue because
it doesn't have a handwritten .types file. Instead, it relies on gtk-doc
generating it, and uses IGNORE_HFILES to filter out internal headers.
This seems more maintainable, so if anyone wants to resurrect this old
documentation, I'd do it that way.