mctest: allow running FakeCM on a separate bus connection
This was possible before, but not exactly straightforward: you had to
clone the guts of create_fakecm_account(). By adding an optional
parameter to create_fakecm_account(), and a method to
SimulatedConnection to drop its bus name, it becomes pretty easy to run
the fake CM on a separate bus, and hence simulate it crashing.