mctest: split up enable_fakecm_account()
This patch splits out the second half of enable_fakecm_account(),
starting at the point where it expects a RequestConnection call, into a
new function, expect_fakecm_connection(). enable_fakecm_account() calls
it, so its behaviour does not change. But the connectivity-monitoring
tests will want to expect a call to RequestConnection() and let that
part of the standard check run to completion, so will be able to use the
freshly-split-out function.