Master: poison singleton pointer when disposed
Previously, when the 'default' instance of McdMaster (i.e., the
singleton) was disposed, the 'default_master' variable was left pointing
to it. This is basically fine in practice, since the McdMaster is meant
to survive for the lifetime of the connection. But I hit a weird crash
(caused by the test suite's version of MC waiting 5 seconds after the
McdMaster is destroyed before closing, which interacts badly with the
reconnection timeout being less than 5 seconds) and it would have been
much clearer what was happening if 'default_master' had held a known-bad
address rather than a possibly-valid pointer to a freed object.

This doesn't change the behaviour of the code in any significant way:
we'll still crash in this situation. But as far as I can tell the bug is
harmless.