mctest: resuscitate MC asynchronously
Shortly, MC will start making blocking calls on the test suite's fake NM
implementation (via nm-glib).

Calling bus.get_object service-activated MC as a side-effect of making a
blocking call to Introspect() on it, so it deadlocked. I think it's
clearer to explicitly, asynchronously activate it, and wait for it to
claim its names; which also avoids the deadlock.