account-manager/create-with-properties: modernize and asyncify
For some reason, the blocking call to Introspect here reliably
deadlocked against the blocking call MC will, in a couple of patches'
time, make to (the fake) Network Manager service implemented by the test
suite. Removing it just made the GetAll() call deadlock instead. If I
render that asynchronous, the deadlock goes away.

I am not overly happy with having to do this kind of ad-hoc kludging of
the test suite to make it run, but we're working around a library bug
and I have no interest in delaying the NM integration while I fix
nm-glib.