Master: Always bind connection to a transport
Previously, when a connection was brought online in response to
an event other than a transport becoming connected, it would always be
bound to a transport (if any transport plugins are loaded, obviously),
regardless of whether the dictionary o' conditions was empty or not.
However, when a transport coming online caused a connection to be
brought up, that connection would not be bound to the transport that
caused it to come online if the conditions were empty (which they
generally are).

For reasons I can't quite fathom, this actually doesn't seem to have
affected anything: it could be because monitor_state_changed_cb in
McdKludgeTransport also binds waiting connections to the transport that
came online. But this seems like a sensible change for sanity and
consistency reasons.