McdClient: don't allow unique_name to become NULL
_mcd_client_registry_found_name() is called in three situations:

• the initial call to ListNames() has returned;
• the initial call to ListActivatableNames() has returned; and
• NameOwnerChanged has been emitted for a client-shaped well-known name.

In the first two cases, unique_name_if_known is NULL; in the first case,
activatable is FALSE.

I don't really understand what business the McdClientRegistry has
telling the McdClient what its unique name is even when it's non-NULL,
but passing NULL to _mcd_client_proxy_set_active() made McdClient
violate an internal invariant. The first action McdClient takes is to
discover the client's unique name, if any; if the client is not running,
unique_name is set to "" (not NULL). Other parts of the McdClient
internals assume (and assert) that unique_name is not NULL by the time
they run.

I *believe* that the client registry need not ever call
_mcd_client_proxy_set_active(), but I am loathe to make that change
without fully understanding the code and writing some tests.

Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=39767>
Reviewed-by: Xavier Claessens <xclaesse@gmail.com>